Change tabbarController

+++ LeanChat/LeanChat/manager/CDUtils.m
@@ -160,7 +160,7 @@
 + (NSString *)uuid {
     NSString *chars = @"abcdefghijklmnopgrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789";
     assert(chars.length == 62);
-    int len = chars.length;
+    int len = (int)chars.length;
     NSMutableString *result = [[NSMutableString alloc] init];
     for (int i = 0; i < 24; i++) {
         int p = arc4random_uniform(len);
